appveyor: do not override clone_script
previously, we overrided the repo cloning method by specifying the
"clone_script" section. but when the "git clone" command failed to
fetch the merge commit of a pull request under test.

in this change, instead of overriding the clone method, we just
reuse the default cloning method, and specify the install section
to prepare the submodules.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>